I tried original J equation. I get consistent sometimes, but no convergence really.
I should make sure to figure out the recursive equations for J, but also make sure that the updates happen in the correct way. I'm suspecting that the downdate is a problem.  I should check the theta estimate when k = dx true but after it has downdated. And maybe I can keep J equation fixed until I find the other error. Another issue is the sampling the weights when they are all too similar. Ask Petar.